Extract shared pricing module
Move menu price normalization helpers into jo.pricing so menu and prompt NUI can share the same price structure without prompt-nui requiring the full menu module.

Keep jo.menu price helpers as compatibility aliases and update internal menu/prompt calls to use jo.pricing directly.
