+++ jo_libs/modules/menu/client.lua
@@ -1,4 +1,5 @@
 jo.menu.exports = {}
+jo.require("pricing")
 jo.require("timeout")
 jo.require("nui")
 jo.require("string")
@@ -228,7 +229,7 @@ local function formatItemPrice(item)
         jo.menu.displayLoader()
         loaderOn = true
       end
-      price = jo.menu.formatPrice(price)
+      price = jo.pricing.formatPrice(price)
       if loaderOn then
         SetTimeout(100, jo.menu.hideLoader)
       end
@@ -368,7 +369,7 @@ function jo.menu.updateItem(id, index, key, value) menus[id]:updateItem(index, k
 function MenuClass:updateValue(keys, value)
   if type(keys) ~= "table" then keys = { keys } end
   if keys[#keys] == "price" then
-    value = jo.menu.formatPrice(value)
+    value = jo.pricing.formatPrice(value)
   end
   local v = table.copy(value)
   table.upsert(self, keys, v)
